pkg/store: add store.Saver interface, add GetLatestBackup()
Issue: https://github.com/adamdecaf/cert-manage/issues/53
